[iOS] clean ups to platform view controller
