+++ shell/platform/darwin/ios/BUILD.gn
@@ -77,7 +77,6 @@ source_set("flutter_framework_source_arc") {
     "framework/Source/FlutterMetalLayer.mm",
     "framework/Source/FlutterOverlayView.h",
     "framework/Source/FlutterOverlayView.mm",
-    "framework/Source/FlutterPlatformViews.mm",
     "framework/Source/FlutterPlatformViews_Internal.h",
     "framework/Source/FlutterPlatformViews_Internal.mm",
     "framework/Source/FlutterPluginAppLifeCycleDelegate.mm",
@@ -112,6 +111,8 @@ source_set("flutter_framework_source_arc") {
     "framework/Source/connection_collection.mm",
     "framework/Source/platform_message_response_darwin.h",
     "framework/Source/platform_message_response_darwin.mm",
+    "framework/Source/platform_views_controller.h",
+    "framework/Source/platform_views_controller.mm",
     "framework/Source/profiler_metrics_ios.h",
     "framework/Source/profiler_metrics_ios.mm",
     "framework/Source/vsync_waiter_ios.h",
