+++ shell/platform/darwin/ios/framework/Source/FlutterView.h
@@ -18,7 +18,7 @@
 - (flutter::Rasterizer::Screenshot)takeScreenshot:(flutter::Rasterizer::ScreenshotType)type
                                   asBase64Encoded:(BOOL)base64Encode;
 
-- (std::shared_ptr<flutter::FlutterPlatformViewsController>&)platformViewsController;
+- (std::shared_ptr<flutter::PlatformViewsController>&)platformViewsController;
 
 /**
  * A callback that is called when iOS queries accessibility information of the Flutter view.
