+++ shell/platform/darwin/ios/framework/Source/accessibility_bridge.h
@@ -51,7 +51,7 @@ class AccessibilityBridge final : public AccessibilityBridgeIos {
 
   AccessibilityBridge(FlutterViewController* view_controller,
                       PlatformViewIOS* platform_view,
-                      std::shared_ptr<FlutterPlatformViewsController> platform_views_controller,
+                      std::shared_ptr<PlatformViewsController> platform_views_controller,
                       std::unique_ptr<IosDelegate> ios_delegate = nullptr);
   ~AccessibilityBridge();
 
@@ -73,7 +73,7 @@ class AccessibilityBridge final : public AccessibilityBridgeIos {
 
   fml::WeakPtr<AccessibilityBridge> GetWeakPtr();
 
-  std::shared_ptr<FlutterPlatformViewsController> GetPlatformViewsController() const override {
+  std::shared_ptr<PlatformViewsController> GetPlatformViewsController() const override {
     return platform_views_controller_;
   };
 
@@ -92,7 +92,7 @@ class AccessibilityBridge final : public AccessibilityBridgeIos {
 
   FlutterViewController* view_controller_;
   PlatformViewIOS* platform_view_;
-  const std::shared_ptr<FlutterPlatformViewsController> platform_views_controller_;
+  const std::shared_ptr<PlatformViewsController> platform_views_controller_;
   // If the this id is kSemanticObjectIdInvalid, it means either nothing has
   // been focused or the focus is currently outside of the flutter application
   // (i.e. the status bar or keyboard)
