+++ shell/platform/darwin/ios/framework/Source/accessibility_bridge_ios.h
@@ -14,7 +14,7 @@
 @class UIView;
 
 namespace flutter {
-class FlutterPlatformViewsController;
+class PlatformViewsController;
 
 /// Interface that represents an accessibility bridge for iOS.
 class AccessibilityBridgeIos {
@@ -39,7 +39,7 @@ class AccessibilityBridgeIos {
    * The input id is the uid of the newly focused SemanticObject.
    */
   virtual void AccessibilityObjectDidLoseFocus(int32_t id) = 0;
-  virtual std::shared_ptr<FlutterPlatformViewsController> GetPlatformViewsController() const = 0;
+  virtual std::shared_ptr<PlatformViewsController> GetPlatformViewsController() const = 0;
 };
 
 }  // namespace flutter
