+++ shell/platform/darwin/ios/framework/Source/accessibility_bridge_test.mm
@@ -279,8 +279,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
                                  /*ui=*/thread_task_runner,
                                  /*io=*/thread_task_runner);
 
-    auto flutterPlatformViewsController =
-        std::make_shared<flutter::FlutterPlatformViewsController>();
+    auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
     flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
     auto platform_view = std::make_unique<flutter::PlatformViewIOS>(
         /*delegate=*/mock_delegate,
@@ -341,8 +340,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
                                  /*ui=*/thread_task_runner,
                                  /*io=*/thread_task_runner);
 
-    auto flutterPlatformViewsController =
-        std::make_shared<flutter::FlutterPlatformViewsController>();
+    auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
     flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
     auto platform_view = std::make_unique<flutter::PlatformViewIOS>(
         /*delegate=*/mock_delegate,
@@ -389,7 +387,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
                                /*ui=*/thread_task_runner,
                                /*io=*/thread_task_runner);
 
-  auto flutterPlatformViewsController = std::make_shared<flutter::FlutterPlatformViewsController>();
+  auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
   flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
   auto platform_view = std::make_unique<flutter::PlatformViewIOS>(
       /*delegate=*/mock_delegate,
@@ -486,7 +484,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
                                /*ui=*/thread_task_runner,
                                /*io=*/thread_task_runner);
 
-  auto flutterPlatformViewsController = std::make_shared<flutter::FlutterPlatformViewsController>();
+  auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
   flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
   auto platform_view = std::make_unique<flutter::PlatformViewIOS>(
       /*delegate=*/mock_delegate,
@@ -561,7 +559,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
                                /*ui=*/thread_task_runner,
                                /*io=*/thread_task_runner);
 
-  auto flutterPlatformViewsController = std::make_shared<flutter::FlutterPlatformViewsController>();
+  auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
   flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
   auto platform_view = std::make_unique<flutter::PlatformViewIOS>(
       /*delegate=*/mock_delegate,
@@ -2181,8 +2179,7 @@ fml::RefPtr<fml::TaskRunner> CreateNewThread(const std::string& name) {
         /*is_gpu_disabled_sync_switch=*/std::make_shared<fml::SyncSwitch>());
 
     id mockFlutterViewController = OCMClassMock([FlutterViewController class]);
-    auto flutterPlatformViewsController =
-        std::make_shared<flutter::FlutterPlatformViewsController>();
+    auto flutterPlatformViewsController = std::make_shared<flutter::PlatformViewsController>();
     flutterPlatformViewsController->SetTaskRunner(thread_task_runner);
 
     OCMStub([mockFlutterViewController platformViewsController])
