+++ shell/platform/darwin/ios/ios_external_view_embedder.h
@@ -12,16 +12,15 @@ namespace flutter {
 
 class IOSExternalViewEmbedder : public ExternalViewEmbedder {
  public:
-  IOSExternalViewEmbedder(const std::shared_ptr<FlutterPlatformViewsController>&
-                              platform_views_controller,
-                          const std::shared_ptr<IOSContext>& context);
+  IOSExternalViewEmbedder(
+      const std::shared_ptr<PlatformViewsController>& platform_views_controller,
+      const std::shared_ptr<IOSContext>& context);
 
   // |ExternalViewEmbedder|
   virtual ~IOSExternalViewEmbedder() override;
 
  private:
-  const std::shared_ptr<FlutterPlatformViewsController>&
-      platform_views_controller_;
+  const std::shared_ptr<PlatformViewsController>& platform_views_controller_;
   std::shared_ptr<IOSContext> ios_context_;
 
   // |ExternalViewEmbedder|
