+++ shell/platform/darwin/ios/ios_external_view_embedder.mm
@@ -12,7 +12,7 @@ FLUTTER_ASSERT_ARC
 namespace flutter {
 
 IOSExternalViewEmbedder::IOSExternalViewEmbedder(
-    const std::shared_ptr<FlutterPlatformViewsController>& platform_views_controller,
+    const std::shared_ptr<PlatformViewsController>& platform_views_controller,
     const std::shared_ptr<IOSContext>& context)
     : platform_views_controller_(platform_views_controller), ios_context_(context) {
   FML_CHECK(ios_context_);
