+++ shell/platform/darwin/ios/platform_view_ios.mm
@@ -42,7 +42,7 @@ void PlatformViewIOS::AccessibilityBridgeManager::Clear() {
 PlatformViewIOS::PlatformViewIOS(
     PlatformView::Delegate& delegate,
     const std::shared_ptr<IOSContext>& context,
-    const std::shared_ptr<FlutterPlatformViewsController>& platform_views_controller,
+    const std::shared_ptr<PlatformViewsController>& platform_views_controller,
     const flutter::TaskRunners& task_runners)
     : PlatformView(delegate, task_runners),
       ios_context_(context),
@@ -54,7 +54,7 @@ PlatformViewIOS::PlatformViewIOS(
 PlatformViewIOS::PlatformViewIOS(
     PlatformView::Delegate& delegate,
     IOSRenderingAPI rendering_api,
-    const std::shared_ptr<FlutterPlatformViewsController>& platform_views_controller,
+    const std::shared_ptr<PlatformViewsController>& platform_views_controller,
     const flutter::TaskRunners& task_runners,
     const std::shared_ptr<fml::ConcurrentTaskRunner>& worker_task_runner,
     const std::shared_ptr<const fml::SyncSwitch>& is_gpu_disabled_sync_switch)
